feat: 채팅방 입장

+++ spring/webchat/src/main/java/com/example/webchat/chat/room/dto/ChatRoomResponse.java
@@ -7,9 +7,9 @@ import lombok.Getter;
 public class ChatRoomResponse {
 
     private final ChatRoom chatRoom;
-    private final long userCount;
+    private final int userCount;
 
-    public ChatRoomResponse(ChatRoom chatRoom, long userCount) {
+    public ChatRoomResponse(ChatRoom chatRoom, int userCount) {
         this.chatRoom = chatRoom;
         this.userCount = userCount;
     }
